+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Genetics and computers - part 1".
 - Tests for "Retro typewriter art".
 - Tests for "Parse SQL queries".
+- Tests for "By train or by car ?".
 
 ### Changed
 - Renaming "Linear Bézier curves" to "Cubic Bézier curves".